digihub-chatbot-data-pipeline/requirements.txt
openai==1.55.3
langchain-community==0.3.0
langchain-core==0.3.21
langchain-openai==0.2.11
#torch==2.2.2
docling==2.52.0
#docling-ibm-models==3.4.3
PyMuPDF==1.25.5
azure-cosmos==4.9.0
azure-identity==1.16.1
python-dotenv==0.21.1
azure.storage.blob==12.16.0
azure-keyvault-secrets==4.9.0
azure-eventhub==5.15.0
requests==2.32.4
fastapi==0.115.12
uvicorn==0.23.2
pytest==8.3.5
psutil==7.0.0
PyPDF2==3.0.1
pydantic==2.11.3
pytest-asyncio==1.1.0